+++ apps/dapp/src/pages/lending/Borrow/index.ts
@@ -1,0 +1,1 @@
+export { default as BorrowDialog } from './BorrowDialog';
