+++ apps/dapp/src/pages/lending/Debt/index.ts
@@ -1,0 +1,1 @@
+export { default as DebtPositions } from './DebtPositions';
