+++ apps/dapp/src/pages/lending/Lend/index.ts
@@ -1,0 +1,2 @@
+export { default as LendDialog } from './LendDialog';
+export { default as LendingPositions } from './LendingPositions';
