+++ apps/dapp/src/pages/lending/Markets/index.ts
@@ -1,0 +1,1 @@
+export { default as Markets } from './Markets';
