+++ apps/dapp/src/pages/lending/Stats/Chart.tsx
@@ -92,3 +92,5 @@ export const Chart = (props: ChartProps) => {
     </Box>
   );
 };
+
+export default Chart;
