+++ apps/dapp/src/pages/lending/Stats/index.ts
@@ -1,0 +1,1 @@
+export { default as Chart } from './Chart';
